VMM/TMR3UtcNow: Log time changes between two calls larger than 30min. [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@69951 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR3/TM.cpp
@@ -2835,7 +2835,7 @@ VMMR3_INT_DECL(PRTTIMESPEC) TMR3UtcNow(PVM pVM, PRTTIMESPEC pTime)
      */
     int64_t nsPrev = ASMAtomicXchgS64(&pVM->tm.s.nsLastUtcNow, RTTimeSpecGetNano(pTime));
     int64_t cNsDelta = RTTimeSpecGetNano(pTime) - nsPrev;
-    if (RT_ABS(cNsDelta) > RT_NS_1HOUR / 2)
+    if ((uint64_t)RT_ABS(cNsDelta) > RT_NS_1HOUR / 2)
         LogRel(("TMR3UtcNow: nsNow=%RI64 nsPrev=%RI64 -> cNsDelta=%RI64 (offLag=%RI64 offVirtualSync=%RU64 offVirtualSyncGivenUp=%RU64)\n",
                 RTTimeSpecGetNano(pTime), nsPrev, cNsDelta, offLag, offVirtualSync, offVirtualSyncGivenUp));
